feat: don't emit gauges that haven't been recorded
Currently, when a `Gauge` or `IntGauge` metric has been registered but
has never had a value recorded, it will be included in a scrape with a
value of 0. This can create incorrect results when a scrape occurs when
a gauge has been registered but its value hasn't been recorded yet.